Use AggregatableTagMappings on TagMigration and TaggingSpreadsheet
