+++ app/models/tag_migration.rb
@@ -1,4 +1,6 @@
 class TagMigration < ActiveRecord::Base
+  include AggregatableTagMappings
+
   has_many :tag_mappings, dependent: :destroy, as: :tagging_source
   validates :source_content_id, presence: true
 
@@ -17,26 +19,5 @@ class TagMigration < ActiveRecord::Base
 
   def error_count
     tag_mappings.errored.count
-  end
-
-  def aggregated_tag_mappings
-    tag_mappings_grouped_by_content_base_path.reduce([]) do |accumulator, aggregation|
-      accumulator << AggregatedTagMapping.new(content_base_path: aggregation.first, tag_mappings: aggregation.last)
-    end
-  end
-
-private
-
-  def tag_mappings_grouped_by_content_base_path
-    tag_mappings.by_state.by_content_base_path.by_link_title
-      .select(
-        :id,
-        :link_type,
-        :link_title,
-        :content_base_path,
-        :messages,
-        :link_content_id,
-        :state)
-      .group_by(&:content_base_path)
   end
 end
